Temporary commit of images feature
Implemented image upload to the server via drag and drop on the canvas
of Board.
Images are saved in the server in the main DB.

Images are sent to each client on connection (the server iterates the
images in the room and sends them one after the other to the client).
On each image upload all the clients in that room get the new image.

Images can be shown via new button added to the board buttons - the
"plus" button which will switch the images' dialog hidden property
on/off.

This is merely a temporary commit since the feature is far from done and
some architectural changes are still necessary (not to mention UI
changes, since it looks horrible)`